# Conflicts:
#	src/Nethermind/Nethermind.Init/Steps/Migrations/ReceiptFixMigration.cs

+++ src/Nethermind/Nethermind.Blockchain.Test/BlockhashProviderTests.cs
@@ -143,7 +143,7 @@ public class BlockhashProviderTests
             current = Build.A.Block.WithParent(current).TestObject;
         }
 
-        long lookupNumber = current.Number - 256;
+        ulong lookupNumber = current.Number - 256ul;
         Hash256? result = provider.GetBlockhash(current.Header, lookupNumber, Frontier.Instance);
         Assert.That(result, Is.Not.Null);
     }
@@ -173,21 +173,21 @@ public class BlockhashProviderTests
     }
 
     [MaxTime(Timeout.MaxTestTime)]
-    [TestCase(512, -1, true, TestName = "Can_get_parent_hash")]
-    [TestCase(512, 0, false, TestName = "Cannot_ask_for_self")]
-    [TestCase(512, 1, false, TestName = "Cannot_ask_about_future")]
-    [TestCase(512, -256, true, TestName = "Can_lookup_up_to_256_before")]
-    [TestCase(512, -257, false, TestName = "No_lookup_more_than_256_before")]
-    public void Blockhash_lookup_with_full_chain(int chainLength, int lookupOffset, bool expectNonNull)
+    [TestCase(512ul, -1, true, TestName = "Can_get_parent_hash")]
+    [TestCase(512ul, 0, false, TestName = "Cannot_ask_for_self")]
+    [TestCase(512ul, 1, false, TestName = "Cannot_ask_about_future")]
+    [TestCase(512ul, -256, true, TestName = "Can_lookup_up_to_256_before")]
+    [TestCase(512ul, -257, false, TestName = "No_lookup_more_than_256_before")]
+    public void Blockhash_lookup_with_full_chain(ulong chainLength, int lookupOffset, bool expectNonNull)
     {
         Block genesis = Build.A.Block.Genesis.TestObject;
         BlockTreeBuilder blockTreeBuilder = Build.A.BlockTree(genesis).OfChainLength(chainLength);
         BlockTree tree = blockTreeBuilder.TestObject;
 
         BlockhashProvider provider = CreateBlockHashProvider(blockTreeBuilder.HeaderStore, Frontier.Instance);
-        BlockHeader head = tree.FindHeader(chainLength - 1, BlockTreeLookupOptions.None)!;
+        BlockHeader head = tree.FindHeader(chainLength - 1ul, BlockTreeLookupOptions.None)!;
         Block current = Build.A.Block.WithParent(head).TestObject;
-        long lookupNumber = chainLength + lookupOffset;
+        ulong lookupNumber = (ulong)((long)chainLength + lookupOffset);
         Hash256? result = provider.GetBlockhash(current.Header, lookupNumber, Frontier.Instance);
 
         if (expectNonNull)
@@ -218,17 +218,17 @@ public class BlockhashProviderTests
     }
 
     [MaxTime(Timeout.MaxTestTime)]
-    [TestCase(1)]
-    [TestCase(512)]
-    [TestCase(8192)]
-    [TestCase(8193)]
-    public void Eip2935_enabled_Eip7709_disabled_and_then_get_hash(int chainLength)
+    [TestCase(1ul)]
+    [TestCase(512ul)]
+    [TestCase(8192ul)]
+    [TestCase(8193ul)]
+    public void Eip2935_enabled_Eip7709_disabled_and_then_get_hash(ulong chainLength)
     {
         Block genesis = Build.A.Block.Genesis.TestObject;
         BlockTreeBuilder blockTreeBuilder = Build.A.BlockTree(genesis).OfHeadersOnly.OfChainLength(chainLength);
         BlockTree tree = blockTreeBuilder.TestObject;
 
-        BlockHeader? head = tree.FindHeader(chainLength - 1, BlockTreeLookupOptions.None);
+        BlockHeader? head = tree.FindHeader(chainLength - 1ul, BlockTreeLookupOptions.None);
         // number = chainLength
 
         (IWorldState worldState, Hash256 stateRoot) = CreateWorldState();
@@ -271,11 +271,11 @@ public class BlockhashProviderTests
     [Test, MaxTime(Timeout.MaxTestTime)]
     public void Eip2935_poc_trimmed_hashes()
     {
-        int chainLength = 42;
+        ulong chainLength = 42ul;
         Block genesis = Build.A.Block.Genesis.TestObject;
         BlockTree tree = Build.A.BlockTree(genesis).OfHeadersOnly.OfChainLength(chainLength).TestObject;
 
-        BlockHeader? head = tree.FindHeader(chainLength - 1, BlockTreeLookupOptions.None);
+        BlockHeader? head = tree.FindHeader(chainLength - 1ul, BlockTreeLookupOptions.None);
         // number = chainLength
 
         (IWorldState worldState, Hash256 stateRoot) = CreateWorldState();
@@ -309,7 +309,7 @@ public class BlockhashProviderTests
         IWorldState balWorldState = CreateWorldStateWithHistoryContract(spec);
         Block parent = Build.A.Block.WithNumber(41).TestObject;
         Block current = Build.A.Block.WithParent(parent).TestObject;
-        UInt256 parentBlockIndex = new((ulong)((current.Number - 1) % spec.Eip2935RingBufferSize));
+        UInt256 parentBlockIndex = new((current.Number - 1) % spec.Eip2935RingBufferSize);
         StorageCell storageCell = new(Eip2935Constants.BlockHashHistoryAddress, parentBlockIndex);
 
         using IDisposable legacyScope = legacyWorldState.BeginScope(current.Header);
@@ -373,7 +373,7 @@ public class BlockhashProviderTests
         // At block 150 with buffer size 100, we need hashes for blocks 50-149
         // Block 150 stores block 149's hash (already done above)
         // Now process blocks 51-149 from the tree to store blocks 50-148
-        for (long blockNum = chainLength - customRingBufferSize + 1; blockNum < chainLength; blockNum++)
+        for (ulong blockNum = chainLength - customRingBufferSize + 1; blockNum < chainLength; blockNum++)
         {
             BlockHeader? header = tree.FindHeader(blockNum, BlockTreeLookupOptions.None);
             Assert.That(header, Is.Not.Null, $"Block {blockNum} should exist in tree");
@@ -383,7 +383,7 @@ public class BlockhashProviderTests
 
         // Now verify all blocks behave correctly with custom ring buffer size
         // At block 150 with buffer size 100, only blocks [50, 149] should be retrievable
-        for (long blockNum = 1; blockNum < chainLength - customRingBufferSize; blockNum++)
+        for (ulong blockNum = 1ul; blockNum < (chainLength - customRingBufferSize); blockNum++)
         {
             Hash256? result = store.GetBlockHashFromState(current.Header, blockNum, customSpec);
 
@@ -391,7 +391,7 @@ public class BlockhashProviderTests
                 $"Block {blockNum} should be outside custom ring buffer of size {customRingBufferSize} (proves custom size is used, not default 8191)");
         }
 
-        for (long blockNum = chainLength - customRingBufferSize; blockNum < chainLength; blockNum++)
+        for (ulong blockNum = chainLength - customRingBufferSize; blockNum < chainLength; blockNum++)
         {
             BlockHeader? expectedHeader = tree.FindHeader(blockNum, BlockTreeLookupOptions.None);
             Hash256? result = store.GetBlockHashFromState(current.Header, blockNum, customSpec);
@@ -414,7 +414,7 @@ public class BlockhashProviderTests
         SlowHeaderStore slowHeaderStore = new(blockTreeBuilder.HeaderStore) { SlowBlockNumber = 2 };
         BlockTree tree = blockTreeBuilder.TestObject;
         BlockHeader head = tree.FindHeader(chainLength - 1, BlockTreeLookupOptions.None)!;
-        long expectedBlockNumber = head.Number - 2;
+        ulong expectedBlockNumber = head.Number - 2;
         BlockHeader expected = tree.FindHeader(expectedBlockNumber, BlockTreeLookupOptions.None)!;
         BlockHeader previousHead = tree.FindHeader(chainLength - 4, BlockTreeLookupOptions.None)!;
         BlockhashProvider provider = CreateBlockHashProvider(slowHeaderStore, Frontier.Instance);
